Code cleanup (#976)
* separated ported into files to more closely match the originals
* fixed some minor analysis issues
