+++ src/Integration.Vsix/CFamily/PortedFromJava/ReadMe.txt
@@ -1,0 +1,7 @@
+﻿All of the files in this folder were ported from the sonar-cfamily-plugin:
+  https://github.com/SonarSource/sonar-cpp/blob/master/sonar-cfamily-plugin
+
+Each file is as direct as translation as possible to make comparing the ported version against
+the original Java version easier (i.e. so we can more easier port modifications in the future).
+
+For this reason, any issues should be suppressed rather than fixed in the ported code.
